removed old analyzer selection dialog
