--- fapolicy_analyzer/glade/analyzer_selection_dialog.glade
@@ -1,138 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright Concurrent Technologies Corporation 2021
-
-  This program is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  This program is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with this program.  If not, see <https://www.gnu.org/licenses/>.
--->
-
-<!-- Generated with glade 3.38.2 -->
-<interface>
-  <requires lib="gtk+" version="3.22"/>
-  <object class="GtkDialog" id="analyzerSelectionDialog">
-    <property name="can-focus">False</property>
-    <property name="modal">True</property>
-    <property name="default-width">320</property>
-    <property name="default-height">160</property>
-    <property name="type-hint">dialog</property>
-    <property name="decorated">False</property>
-    <child internal-child="vbox">
-      <object class="GtkBox">
-        <property name="can-focus">False</property>
-        <property name="orientation">vertical</property>
-        <property name="spacing">2</property>
-        <property name="homogeneous">True</property>
-        <child internal-child="action_area">
-          <object class="GtkButtonBox">
-            <property name="can-focus">False</property>
-            <property name="layout-style">end</property>
-            <child>
-              <object class="GtkButton" id="scanSystemBtn">
-                <property name="label" translatable="yes">Scan System</property>
-                <property name="visible">True</property>
-                <property name="sensitive">False</property>
-                <property name="can-focus">True</property>
-                <property name="receives-default">True</property>
-              </object>
-              <packing>
-                <property name="expand">False</property>
-                <property name="fill">True</property>
-                <property name="position">0</property>
-              </packing>
-            </child>
-            <child>
-              <object class="GtkButton" id="adminTrustDatabasesBtn">
-                <property name="label" translatable="yes">Administer Trust Databases</property>
-                <property name="visible">True</property>
-                <property name="can-focus">True</property>
-                <property name="receives-default">True</property>
-              </object>
-              <packing>
-                <property name="expand">False</property>
-                <property name="fill">True</property>
-                <property name="position">1</property>
-              </packing>
-            </child>
-          </object>
-          <packing>
-            <property name="expand">True</property>
-            <property name="fill">True</property>
-            <property name="position">0</property>
-          </packing>
-        </child>
-        <child>
-          <object class="GtkLabel">
-            <property name="visible">True</property>
-            <property name="can-focus">False</property>
-            <property name="label" translatable="yes">Analyzer Selection</property>
-            <attributes>
-              <attribute name="weight" value="ultrabold"/>
-            </attributes>
-          </object>
-          <packing>
-            <property name="expand">False</property>
-            <property name="fill">True</property>
-            <property name="position">1</property>
-          </packing>
-        </child>
-        <child>
-          <object class="GtkBox">
-            <property name="visible">True</property>
-            <property name="can-focus">False</property>
-            <property name="halign">center</property>
-            <property name="valign">center</property>
-            <property name="spacing">10</property>
-            <child>
-              <object class="GtkEntry" id="auditLogTxt">
-                <property name="visible">True</property>
-                <property name="can-focus">True</property>
-                <property name="secondary-icon-stock">gtk-open</property>
-                <signal name="changed" handler="on_auditLogTxt_changed" swapped="no"/>
-                <signal name="icon-press" handler="on_auditLogTxt_icon_press" swapped="no"/>
-              </object>
-              <packing>
-                <property name="expand">False</property>
-                <property name="fill">True</property>
-                <property name="position">0</property>
-              </packing>
-            </child>
-            <child>
-              <object class="GtkButton" id="analyzeAuditBtn">
-                <property name="label" translatable="yes">Analyze From Audit</property>
-                <property name="visible">True</property>
-                <property name="sensitive">False</property>
-                <property name="can-focus">True</property>
-                <property name="receives-default">True</property>
-              </object>
-              <packing>
-                <property name="expand">False</property>
-                <property name="fill">True</property>
-                <property name="position">1</property>
-              </packing>
-            </child>
-          </object>
-          <packing>
-            <property name="expand">False</property>
-            <property name="fill">True</property>
-            <property name="position">2</property>
-          </packing>
-        </child>
-      </object>
-    </child>
-    <action-widgets>
-      <action-widget response="2">analyzeAuditBtn</action-widget>
-      <action-widget response="1">scanSystemBtn</action-widget>
-      <action-widget response="0">adminTrustDatabasesBtn</action-widget>
-    </action-widgets>
-  </object>
-</interface>
